+++ python/paddle/distributed/__init__.py
@@ -72,7 +72,7 @@ from .entry_attr import ShowClickEntry  # noqa: F401
 # (TODO: GhostScreaming) It needs migration of ParallelEnv. However,
 # it's hard to migrate APIs in paddle.fluid.dygraph.parallel completely.
 # It will be replaced later.
-from paddle.fluid.dygraph.parallel import ParallelEnv  # noqa: F401
+from .parallel import ParallelEnv  # noqa: F401
 
 from . import cloud_utils  # noqa: F401
