+++ python/paddle/fluid/tests/unittests/auto_parallel/test_fused_linear_pass.py
@@ -21,7 +21,6 @@ from get_gpt_model import FakeDataset, generate_model
 
 import paddle
 from paddle.distributed.fleet import auto
-from paddle.fluid.dygraph.parallel import ParallelEnv
 
 sys.path.append("..")
 from test_sparse_addmm_op import get_cuda_version
@@ -55,7 +54,7 @@ class TestFusedLinearPass(unittest.TestCase):
         paddle.seed(2021)
         np.random.seed(2021)
         random.seed(2021)
-        place = paddle.fluid.CUDAPlace(ParallelEnv().dev_id)
+        place = paddle.fluid.CUDAPlace(paddle.distributed.ParallelEnv().dev_id)
         engine._executor = paddle.static.Executor(place)
 
     def get_engine(self, use_fused_passes=False, fused_passes_list=[]):
